requirements.txt
alabaster==0.7.8
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.12.1
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
Babel==2.6.0
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==3.3.1
blinker==1.4
blobconverter==0.0.10
boto3==1.18.3
botocore==1.21.3
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.3.0
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
depthai==2.7.2.0
distro==1.4.0
Django==2.2.13
django-crispy-forms==1.7.2
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
enum34==1.1.10
feedparser==5.2.1
ffmpy3==0.2.4
gitsome==0.8.0
grpcio==1.16.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
imagesize==1.2.0
IMDbPY==6.8
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.10.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==18.0.1
kiwisolver==1.3.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nemo-emblems==4.8.1
nest-asyncio==1.5.1
netifaces==0.10.4
nose==1.3.7
notebook==6.4.0
numpy==1.21.0
numpydoc==0.7.0
oauthlib==3.1.0
onboard==1.4.1
open3d==0.10.0.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==20.3
PAM==0.4.2
pandas==1.3.0
pandas-ml==0.6.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.9.0
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyRFC3339==1.1
pyrsistent==0.18.0
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-xapp==2.0.2
python-xlib==0.23
pytube==10.8.5
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
reportlab==3.5.34
requests==2.24.0
requests-file==1.4.3
requests-unixsocket==0.2.0
roman==2.0.0
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.0
SecretStorage==2.3.1
Send2Trash==1.7.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sklearn-pandas==2.2.0
soupsieve==1.9.5
Sphinx==1.8.5
sqlparse==0.4.1
systemd-python==234
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tinycss2==1.0.2
tldextract==2.2.1
tornado==6.1
traitlets==5.0.5
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
xkit==0.0.0
xonsh==0.9.13
zeroconf==0.24.4
